wifi: iwlwifi: mvm: remove STARTING state
While removing the retry loops, I neglected to also remove the
bits of code that were added with the retry loops that prevented
firmware restart, now that the retry loops no longer exist they
need to also be removed so that restart/reprobe can happen. Fix
that by removing the STARTING state and associated code.

Fixes: dfdfe4be183b ("wifi: iwlwifi: remove retry loops in start")
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
Signed-off-by: Miri Korenblit <miriam.rachel.korenblit@intel.com>